HPCC-31293 Move thread context functions from jlog to jthread
