+++ system/jlib/jlog.ipp
@@ -1003,8 +1003,4 @@ private:
     }
 };
 
-// Reset logging-related thread-local variables, when a threadpool starts
-extern void resetThreadLogging(const IContextLogger *_logctx, TraceFlags _traceFlags);
-extern void getThreadLoggingInfo(const IContextLogger * &_logctx, TraceFlags &_traceFlags);
-
 #endif
