+++ system/jlib/jtrace.cpp
@@ -49,6 +49,8 @@
 #include "lnuid.h"
 #include <variant>
 
+#include "jthread.hpp"
+
 namespace context     = opentelemetry::context;
 namespace nostd       = opentelemetry::nostd;
 namespace opentel_trace = opentelemetry::trace;
