+++ system/jlib/jtrace.hpp
@@ -332,17 +332,4 @@ extern jlib_decl TraceFlags queryDefaultTraceFlags();
 
 extern jlib_decl TraceFlags loadTraceFlags(const IPropertyTree * globals, const std::initializer_list<TraceOption> & y, TraceFlags dft);
 
-// Temporarily modify the trace context and/or flags for the current thread, for the lifetime of the LogContextScope object
-interface IContextLogger;
-class jlib_decl LogContextScope
-{
-public:
-    LogContextScope(const IContextLogger *ctx);
-    LogContextScope(const IContextLogger *ctx, TraceFlags traceFlags);
-    ~LogContextScope();
-
-    const IContextLogger *prev;
-    TraceFlags prevFlags;
-};
-
 #endif
